+++ src/Step/Tdt/Acte/ChoiceAction/TdtChoiceTypologieActes.php
@@ -1,9 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
+namespace Pastell\Step\Tdt\Acte\ChoiceAction;
+
+use ConnecteurTypeChoiceActionExecutor;
+use Exception;
+use NotFoundException;
+use Pastell\Step\Tdt\Acte\lib\ActesTypePJ;
+use Pastell\Step\Tdt\Acte\lib\ActesTypePJData;
+use TdtConnecteur;
+use UnrecoverableException;
+
 class TdtChoiceTypologieActes extends ConnecteurTypeChoiceActionExecutor
 {
     /**
-     *
      * arrete: arrete
      * autre_document_attache: autre_document_attache
      * type_acte: type_acte
@@ -19,7 +30,7 @@ class TdtChoiceTypologieActes extends ConnecteurTypeChoiceActionExecutor
     /**
      * @throws Exception
      */
-    public function display()
+    public function display(): void
     {
         $connecteur_type_action = $this->getMappingList();
 
@@ -50,7 +61,7 @@ class TdtChoiceTypologieActes extends ConnecteurTypeChoiceActionExecutor
      * @throws Exception
      * @throws UnrecoverableException
      */
-    public function displayAPI()
+    public function displayAPI(): array
     {
         $result = [];
 
@@ -62,25 +73,29 @@ class TdtChoiceTypologieActes extends ConnecteurTypeChoiceActionExecutor
             $this->type,
             TdtConnecteur::FAMILLE_CONNECTEUR
         );
-        if (! $id_ce) {
+        if (!$id_ce) {
             $result['pieces'] = $this->getAllPieces();
             return $result;
         }
 
         $configTdt = $this->getConnecteurConfigByType(TdtConnecteur::FAMILLE_CONNECTEUR);
-        $actesTypePJData->classification_file_path = $configTdt->getFilePath($connecteur_type_action['classification_file'] ?? 'classification_file');
+        $actesTypePJData->classificationFilePath =
+            $configTdt->getFilePath($connecteur_type_action['classification_file'] ?? 'classification_file');
 
-        if (! file_exists($actesTypePJData->classification_file_path)) {
+        if (!file_exists($actesTypePJData->classificationFilePath)) {
             throw new UnrecoverableException("Aucun fichier de classification n'est présent sur le connecteur TDT");
         }
 
-        $actesTypePJData->acte_nature = $this->getDonneesFormulaire()->get($connecteur_type_action['acte_nature'] ?? 'acte_nature');
+        $actesTypePJData->acteNature =
+            $this->getDonneesFormulaire()->get($connecteur_type_action['acte_nature'] ?? 'acte_nature');
 
         $actesTypePJ = $this->objectInstancier->getInstance(ActesTypePJ::class);
 
         $result['actes_type_pj_list'] = $actesTypePJ->getTypePJListe($actesTypePJData);
-        if (! $result['actes_type_pj_list']) {
-            throw new UnrecoverableException("Aucun type de pièce ne correspond pour la nature et la classification selectionnée");
+        if (!$result['actes_type_pj_list']) {
+            throw new UnrecoverableException(
+                'Aucun type de pièce ne correspond pour la nature et la classification sélectionnée'
+            );
         }
 
         $result['pieces'] = $this->getAllPieces();
@@ -88,66 +103,75 @@ class TdtChoiceTypologieActes extends ConnecteurTypeChoiceActionExecutor
     }
 
     /**
-     * @return array|string
      * @throws UnrecoverableException
+     * @throws NotFoundException
      */
-    private function getAllPieces()
+    private function getAllPieces(): array|string
     {
         $connecteur_type_action = $this->getMappingList();
 
         $pieces_list = $this->getDonneesFormulaire()->get($connecteur_type_action['arrete'] ?? 'arrete');
-        if (! $pieces_list) {
+        if (!$pieces_list) {
             throw new UnrecoverableException("La pièce principale n'est pas présente");
         }
-        if ($this->getDonneesFormulaire()->get($connecteur_type_action['autre_document_attache'] ?? 'autre_document_attache')) {
-            $pieces_list = array_merge($pieces_list, $this->getDonneesFormulaire()->get($connecteur_type_action['autre_document_attache'] ?? 'autre_document_attache'));
+        $piecesAnnexe = $this->getDonneesFormulaire()
+            ->get($connecteur_type_action['autre_document_attache'] ?? 'autre_document_attache');
+        if ($piecesAnnexe) {
+            $pieces_list = array_merge($pieces_list, $piecesAnnexe);
         }
         return $pieces_list;
     }
 
-    private function getMappingList()
+    private function getMappingList(): array|bool
     {
         return $this->getDocumentType()->getAction()->getProperties($this->action, 'connecteur-type-mapping');
     }
 
     /**
-     * @return bool
      * @throws Exception
      */
-    public function go()
+    public function go(): bool
     {
 
         $result = [];
-
-
-
 
         $connecteur_type_action = $this->getMappingList();
 
         $type_pj = $this->getRecuperateur()->get('type_pj');
 
         if ((empty($type_pj)) || (!is_array($type_pj))) {
-            throw new UnrecoverableException("Aucun tableau type_pj fourni");
+            throw new UnrecoverableException('Aucun tableau type_pj fourni');
         }
 
         $info = $this->displayAPI();
 
         if ((count($type_pj)) !== (count($info['pieces']))) {
-            throw new UnrecoverableException("Le nombre de type_pj fourni «" . count($type_pj) . "» ne correspond pas au nombre de documents (acte et annexes) «" . (count($info['pieces'])) . "»");
+            throw new UnrecoverableException(
+                sprintf(
+                    'Le nombre de type_pj fourni «%d» ne correspond pas au nombre de documents (acte et annexes) «%d»',
+                    count($type_pj),
+                    count($info['pieces']),
+                )
+            );
         }
         foreach ($type_pj as $i => $type) {
             if (isset($info['actes_type_pj_list']) && !array_key_exists($type, $info['actes_type_pj_list'])) {
-                throw new UnrecoverableException("Le type_pj «" . $type . "» ne correspond pas pour la nature et la classification sélectionnée");
+                throw new UnrecoverableException(
+                    sprintf(
+                        'Le type_pj «%s» ne correspond pas pour la nature et la classification sélectionnée',
+                        $type,
+                    )
+                );
             }
-            $result[] = ['filename' => $info['pieces'][$i], "typologie" => $info['actes_type_pj_list'][$type] ?? $type];
+            $result[] = ['filename' => $info['pieces'][$i], 'typologie' => $info['actes_type_pj_list'][$type] ?? $type];
         }
 
         $this->getDonneesFormulaire()->setData(
             $connecteur_type_action['type_piece'] ?? 'type_piece',
-            count($type_pj) . " fichier(s) typé(s)"
+            count($type_pj) . ' fichier(s) typé(s)'
         );
 
-        $type_acte  = array_shift($type_pj);
+        $type_acte = array_shift($type_pj);
         $this->getDonneesFormulaire()->setData($connecteur_type_action['type_acte'] ?? 'type_acte', $type_acte);
         $this->getDonneesFormulaire()->setData($connecteur_type_action['type_pj'] ?? 'type_pj', json_encode($type_pj));