+++ src/Step/Tdt/Acte/lib/ActesTypePJData.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pastell\Step\Tdt\Acte\lib;
+
+class ActesTypePJData
+{
+    public string $classificationFilePath;
+
+    public string $acteNature;
+}